+++ internal/servers/source.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 
 	"github.com/cloudquery/plugin-sdk/internal/pb"
+	"github.com/cloudquery/plugin-sdk/internal/versions"
 	"github.com/cloudquery/plugin-sdk/plugins"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/specs"
@@ -20,6 +21,12 @@ type SourceServer struct {
 	Plugin  *plugins.SourcePlugin
 	Logger  zerolog.Logger
 	summary *schema.SyncSummary
+}
+
+func (*SourceServer) GetProtocolVersion(context.Context, *pb.GetProtocolVersion_Request) (*pb.GetProtocolVersion_Response, error) {
+	return &pb.GetProtocolVersion_Response{
+		Version: versions.SourceProtocolVersion,
+	}, nil
 }
 
 func (s *SourceServer) GetTables(context.Context, *pb.GetTables_Request) (*pb.GetTables_Response, error) {
@@ -61,7 +68,7 @@ func (s *SourceServer) Sync(req *pb.Sync_Request, stream pb.Source_SyncServer) e
 	var spec specs.Source
 	dec := json.NewDecoder(bytes.NewReader(req.Spec))
 	dec.UseNumber()
-	dec.DisallowUnknownFields()
+	// TODO: warn about unknown fields
 	if err := dec.Decode(&spec); err != nil {
 		return status.Errorf(codes.InvalidArgument, "failed to decode spec: %v", err)
 	}