+++ schema/table.go
@@ -206,6 +206,7 @@ func (t *Table) Resolve(ctx context.Context, meta ClientMeta, parent *Resource, 
 				}
 				wg.Add(1)
 				go func() {
+					defer resourcesSem.Release(1)
 					defer wg.Done()
 					//nolint:all
 					summary.Merge(t.resolveObject(ctx, meta, parent, objects[i], resolvedResources))